requirement.txt
mysql-connector-python==9.0.0
prettytable==3.11.0